backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-groq
chromadb
sentence-transformers
pypdf
pdfplumber
pandas
numpy
scikit-learn
python-dotenv
torch
transformers
langchain-huggingface
langchain-text-splitters
